@@ -8,6 +8,7 @@ import 'package:kifu_viewer/localizations.dart';
 import 'package:kifu_viewer/widgets/home_screen/kifu_viewer.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:menubar/menubar.dart';
+import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shogi/shogi.dart';
 
 class HomeScreen extends StatefulWidget {
@@ -16,9 +17,9 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-  Game _game;
+  Game? _game;
 
-  bool get _hasGame => _game != null && _game.gameBoards.isNotEmpty;
+  bool get _hasGame => _game != null && _game!.gameBoards.isNotEmpty;
 
   bool get _useCustomFont => kIsWeb || Platform.isLinux;
 
@@ -75,10 +76,10 @@ class _HomeScreenState extends State<HomeScreen> {
       body: Center(
         child: _hasGame
             ? DefaultTextStyle(
-                style: Theme.of(context).textTheme.bodyText2.apply(
+                style: Theme.of(context).textTheme.bodyText2!.apply(
                       fontFamily: _useCustomFont ? 'NotoSansJP' : null,
                     ),
-                child: KifuViewer(game: _game),
+                child: KifuViewer(game: _game!),
               )
             : Text(AppLocalizations.homeScreenNoKifu),
       ),
@@ -107,13 +108,13 @@ class _HomeScreenState extends State<HomeScreen> {
   Future<void> _fromClipboard() async {
     final data = await Clipboard.getData('text/plain');
     if (data != null) {
-      _convertFile(data.text);
+      _convertFile(data.text!);
     }
   }
 
   void _convertFile(String file) {
-    final game = Game.fromKif(file);
-    if (game != null) {
+    try {
+      final game = Game.fromKif(file);
       if (game != _game) {
         setState(() => _game = game);
       } else {
@@ -122,7 +123,8 @@ class _HomeScreenState extends State<HomeScreen> {
           description: AppLocalizations.gameAlreadyOpenPopupDescription,
         );
       }
-    } else {
+      // ignore: avoid_catching_errors
+    } on ArgumentError catch (_) {
       _showInvalidContent();
     }
   }
@@ -133,8 +135,8 @@ class _HomeScreenState extends State<HomeScreen> {
       );
 
   void _showCustomDialog({
-    @required String title,
-    @required String description,
+    required String title,
+    required String description,
   }) =>
       showDialog(
         context: context,
@@ -153,11 +155,12 @@ class _HomeScreenState extends State<HomeScreen> {
         ),
       );
 
-  void _showAboutPage() {
+  void _showAboutPage() async {
+    final packageInfo = await PackageInfo.fromPlatform();
     showLicensePage(
       context: context,
       applicationName: 'Kifu Viewer',
-      applicationVersion: '0.2.0+dev1', //TODO package_info not compatible linux, windows
+      applicationVersion: packageInfo.version,
       applicationLegalese: 'defuncart',
     );
   }
